requirements.txt
# Function dependencies, for example:
# package>=version
dialogflow
google-cloud-bigquery
google-cloud-firestore
google-cloud-pubsub
requests
twilio
PyCryptodome
